refactor: make audit scripts clearer
This moves the audit scripts we have so far to their own subfolder in
scripts, with the idea of better distinguishing them and laying the
ground for more like this. A simpler interface via just has also been
added for these.
